Rip out dataspecs everywhere.
