+++ src/fe-analysedriveresponse.cc
@@ -1,7 +1,7 @@
 #include "globals.h"
 #include "flags.h"
 #include "usb/usb.h"
-#include "dataspec.h"
+#include "bitmap.h"
 #include "fluxmap.h"
 #include "decoders/fluxmapreader.h"
 #include "writer.h"
@@ -53,10 +53,20 @@ static StringFlag writeCsv(
 	"Write detailed CSV data",
 	"");
 
-static DataSpecFlag writeImg(
+static StringFlag writeImg(
 	{ "--write-img" },
 	"Draw a graph of the response data",
-	":w=640:h=480");
+	"analysis.png");
+
+static IntFlag imgWidth(
+	{ "--width" },
+	"Width of output graph",
+	800);
+
+static IntFlag imgHeight(
+	{ "--height" },
+	"Height of output graph",
+	600);
 
 /* This is the Turbo colourmap.
  * https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html
@@ -286,16 +296,16 @@ int mainAnalyseDriveResponse(int argc, const char* argv[])
 		}
 	}
 
-	BitmapSpec bitmapSpec(writeImg);
-	if (!bitmapSpec.filename.empty())
+	Bitmap bitmap(writeImg, imgWidth, imgHeight);
+	if (!bitmap.filename.empty())
 	{
-		Agg2D& painter = bitmapSpec.painter();
+		Agg2D& painter = bitmap.painter();
 		painter.clearAll(0xdd, 0xdd, 0xdd);
 
 		const double MARGIN = 30;
 		agg::rect_d drawableBounds = {
 			MARGIN*1.5, MARGIN,
-			bitmapSpec.width - MARGIN, bitmapSpec.height - MARGIN
+			bitmap.width - MARGIN, bitmap.height - MARGIN
 		};
 		agg::rect_d colourbarBounds = {
 			drawableBounds.x2 - MARGIN, drawableBounds.y1,
@@ -347,7 +357,7 @@ int mainAnalyseDriveResponse(int argc, const char* argv[])
 		painter.lineColor(0, 0, 0);
 		painter.rectangle(graphBounds.x1, graphBounds.y1, graphBounds.x2, graphBounds.y2);
 		painter.rectangle(colourbarBounds.x1, drawableBounds.y1, drawableBounds.x2, drawableBounds.y2);
-		bitmapSpec.save();
+		bitmap.save();
 	}
 
     return 0;
