+++ src/fe-read.cc
@@ -8,7 +8,6 @@
 #include "sectorset.h"
 #include "record.h"
 #include "proto.h"
-#include "dataspec.h"
 #include "fluxsource/fluxsource.h"
 #include "fluxsink/fluxsink.h"
 #include "arch/brother/brother.h"
